+++ UnitTests~/AnimationServices/VirtualControllerContextTests.cs
@@ -1,0 +1,11 @@
+﻿using System.Linq;
+using nadena.dev.ndmf.animator;
+using NUnit.Framework;
+using UnityEditor.Animations;
+
+namespace UnitTests.AnimationServices
+{
+    public class VirtualControllerContextTests : TestBase
+    {
+    }
+}